working on basket & discount

+++ yii2-base/common/src/accounting/models/BaseDiscountUsageModelTrait.php
@@ -66,7 +66,7 @@ trait BaseDiscountUsageModelTrait
         enuColumnInfo::selectable => true,
       ],
       'dscusgAmount' => [
-        enuColumnInfo::type       => 'integer',
+        enuColumnInfo::type       => 'double',
         enuColumnInfo::validator  => null,
         enuColumnInfo::default    => null,
         enuColumnInfo::required   => true,
